% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{as.hex}
\alias{as.hex}
\title{Convert RGB or HSL to HEX}
\usage{
as.hex(x)
}
\arguments{
\item{x}{An integer or numeric vector with 3 elements.
An integer vector is assumed to be a value in the RGB color model;
a numeric vector is assumed to be a value in the HSL color model.}
}
\value{
A character string, the hexadecimal code for a value in the RGB
  color model. A hexadecimal code has the form: \code{'#RRGGBB'}, where the
  substrings \code{'RR'}, \code{'GG'}, and \code{'BB'} denote the values of
  the red, green, and blue beams, respectively, in the RGB color model.
}
\description{
\code{as.hex()} converts a value in the RGB (red-green-blue) or
HSL (hue-saturation-lightness) color model to a hexadecimal code that can
be reliably recognized by various plotting functions
(e.g., base graphics or ggplot).
}
\details{
If \code{x} is an integer vector, then it is assumed that \code{x} denotes
a value in the RGB color model---i.e., that the elements of \code{x} are
integers in between 0 and 255 that specify the values of the red, green,
and blue beams, respectively.

If \code{x} is a numeric vector (but not an integer vector), then it is
assumed that \code{x} denotes a value in the HSL color model---i.e.,
that the first element of \code{x} is an whole number between 0 and 359,
denoting the angle of the hue on the color wheel;
that the second element is a floating point number (to 2 decimal places)
between 0 and 1, denoting the saturation proportion;
that the third element is a floating point number (to 2 decimal places)
between 0 and 1, denoting the lightness proportion.

\code{as.hex()} is not vectorized; hence, \code{x} should denote a single
value in either the RGB or HSL color models.
}

