% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{as.hsl}
\alias{as.hsl}
\title{Convert HEX or RGB to HSL}
\usage{
as.hsl(x)
}
\arguments{
\item{x}{A character string or an integer vector with 3 elements.
A character string is assumed to be a hexadecimal code;
an integer vector is assumed to be a value in the RGB color model, whose
elements are integers between 0 and 255}
}
\value{
A numeric vector with three elements.
  The first element denotes the color's hue and is represented as a degree
  on the color wheel (i.e., a whole number between 0 and 359).
  The second element denotes the color's (proportion of) saturation to 2
  decimal places (i.e., a number between 0 and 1.
  The third element denotes the color's (proportion of) lightness to 2
  decimal places (i.e., a number between 0 and 1).
}
\description{
\code{as.hsl()} converts a value in the RGB (red-green-blue) color model to
one in the HSL (hue-saturation-lightness) color model. The value in the RGB
color model may be specified either as a hexadecimal code or as integer
levels of the red, green, and blue beams.
}
\details{
If \code{x} is a character string, then it is assumed that \code{x} is
a hexadecimal code (\code{'#RRGGBB'}) whose substrings denote the the values
of the red (\code{'RR'}), green (\code{'GG'}), and blue (\code{'BB'}) beams.

If \code{x} is an integer vector, then it is assumed that \code{x} denotes
the integer values of the red, green, and blue beams for the color.
}

