% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{as.rgb}
\alias{as.rgb}
\title{Convert HSL or HEX to RGB}
\usage{
as.rgb(x)
}
\arguments{
\item{x}{A character string or a numeric vector with 3 elements.
A character string is assumed to be the hexadecimal code for a value in
the RGB color model;
a numeric vector is assumed to be a value in the HSL color model.}
}
\value{
An integer vector with three elements (valued between 0 and 255) that
  denote the values of the red, green, and blue beams, respectively.
}
\description{
\code{as.rgb()} converts a value in the HSL (hue-saturation-lightness)
color model or a hexadecimal code for a value in the RGB (red-green-blue)
color model to a value in the RGB color model.
}
\details{
If \code{x} is a character string, then it is assumed that \code{x} is
a hexadecimal code (\code{'#RRGGBB'}) whose substrings denote the the values
of the red (\code{'RR'}), green (\code{'GG'}), and blue (\code{'BB'}) beams.

If \code{x} is a numeric vector (but not an integer vector), then it is
assumed that \code{x} denotes a value in the HSL color model---i.e.,
that the first element of \code{x} is an whole number between 0 and 359,
denoting the angle of the hue on the color wheel;
that the second element is a floating point number (to 2 decimal places)
between 0 and 1, denoting the saturation proportion;
that the third element is a floating point number (to 2 decimal places)
between 0 and 1, denoting the lightness proportion.
}

