% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callier.R
\docType{package}
\name{callier}
\alias{callier}
\alias{callier-package}
\title{callier: Color Palettes and Scales for the Callier Center}
\description{
An implementation of the color schemes set out by the Graphic Standards for
the Callier Center for Communication Disorders.
}
\section{License}{

Do not distrubte outside of faculty and staff of the Callier Center
for Communication Disorders.
}

\section{Constants}{

solarOrange

spaceBlue

callierGray

sparkOrange

stratosBlue

skyBlue
}

\section{Color-model converters (internal)}{

as.hex()

as.hsl()

as.rgb()
}

\section{Palette constructors}{

CallierQualitative()

CallierSequential()

CallierDiverging()
}

\section{Scale constructors (for ggplot objects)}{

scale_color_callier()
}

\section{Other functions}{

ColorChart()
}

