% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{CallierDiverging}
\alias{CallierDiverging}
\title{Create Diverging Color Palettes with Callier Colors}
\usage{
CallierDiverging(steps, lower = "orange", lightest = 0.85)
}
\arguments{
\item{steps}{An integer, the number of colors in the palette.}

\item{lower}{A character string that must be matched by either \code{'orange'}
or \code{'blue'}. Default is \code{'orange'}.}

\item{lightest}{A numeric between 0 and 1 that denotes the lightness of the
lightest tint in the middle of the palatte. Values closer to 0 are darker;
those closer to 1 are lighter. Default is 0.85.}
}
\value{
A character vector of hexadecimal codes for colors in the
  RGB (red-green-blue) color model.
}
\description{
\code{CallierDiverging()} creates a diverging color palette. The values on
one half of the palette are tints of \code{solarOrange}; the values on
the other half of the palette are tints of \code{spaceBlue}. The endpoints
of the palette are \code{solarOrange} and \code{spaceBlue}. Colors closer
to the middle of the palette increase in lightness from the endpoint.
}
\details{
A diverging color palette emphasizes: (1) critical values near the middle of
the data range and (2) extreme values at both ends of the data range. This
is accomplished by concatenating two sequential palettes of different hues;
hence, differences above or below the mid-range critical value are
represented by a change in hue and differences within either half of the
palette are represented by changes in lightness within a fixed hue.

If the argument \code{lower} is matched by \code{'orange'}, then the lowest
value of the diverging palette is \code{solarOrange}, and the palette
progresses through lightening orange tints, at the low end, and through
darkening blue tints, at the high end. If \code{lower} is matched by
\code{'blue'}, then the lowest value of the palette is \code{spaceBlue}, and
the palette progresses through lightening blue tints, at the low end, and
through darkening orange tints, at the high end.

Give some consideration to the type of data that are being represented on
the diverging palette and how the structure of that data relates to physical
properties of blue and orange hues. (Blue hues have a shorter wavelength
and a higher frequency than orange hues.) For example, if the data
represented on the diverging palette are (audible) frequencies, then a
palette with orange hues on the \code{lower} end will parallel the order of
the visible frequency scale. Conversely, if the data are distances, then a
palette with blue hues on the \code{lower} end will parallel the order of
the visible wavelength scale.

If the argument \code{steps} is an odd integer, then the color in the middle
of the palette is the \code{lightest} tint of \code{callierGray}.
}
\seealso{
\code{\link{CallierQualitative}}, \code{\link{CallierSequential}}
}

