% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{callierGray}
\alias{#766A62}
\alias{#766a62}
\alias{callierGray}
\title{Callier Gray}
\format{A character string, \code{'#766A62'}.}
\source{
Callier Center Graphic Standards Manual,
  UT Dallas Brand Standards (\url{https://www.utdallas.edu/brand/palette/})
}
\usage{
callierGray
}
\description{
The hexadecimal code that specifies Pantone warm gray 10 ('callier gray'),
which corresponds to 118-106-98 in the RGB (red-green-blue) color model.
}
\details{
Callier gray is one of the three primary Callier Center colors
(solar orange, space blue, callier gray), which should be used when depicting
Callier Center colors in print or electronic formats. Of the primary
Callier Center colors, solar orange has first precedence, space blue has
second, and callier gray has third. Hence, one-color images should use
solar orange; two-color images should use solar orange and space blue; and
three-color images should use solar orange, space blue, and callier gray.
}
\keyword{datasets}

