% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ecoGreen}
\alias{#008542}
\alias{ecoGreen}
\title{Eco Green}
\format{A character string, \code{'#008542'}.}
\source{
Callier Center Graphic Standards Manual,
  UT Dallas Brand Standards (\url{https://www.utdallas.edu/brand/palette/})
}
\usage{
ecoGreen
}
\description{
The hexadecimal code that specifies Pantone 348 ('eco green'), which
corresponds to 0-133-66 in the RGB (red-green-blue) color model.
}
\details{
Eco green is one of the three tertiary Callier Center colors
(eco green, sapling green, seedling green) that are preferred as accents to
complement the primary Callier Center colors in print or electronic formats.
The tertiary Callier Center colors are the three green-hued colors from the
UT Dallas brand color palette.
}
\keyword{datasets}

