% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standards-colors.R
\name{saplingGreen}
\alias{#69BE28}
\alias{#69be28}
\alias{saplingGreen}
\title{Sapling Green}
\source{
Callier Center Graphic Standards Manual,
  UT Dallas Brand Standards (\url{https://www.utdallas.edu/brand/palette/})
}
\usage{
saplingGreen()
}
\value{
A character string, \code{'#69BE28'}.
}
\description{
The hexadecimal code that specifies Pantone 368 ('sapling green'), which
corresponds to 105-190-40 in the RGB (red-green-blue) color model.
}
\details{
Sapling green is one of the three tertiary Callier Center colors
(eco green, sapling green, seedling green) that are preferred as accents to
complement the primary Callier Center colors in print or electronic formats.
The tertiary Callier Center colors are the three green-hued colors from the
UT Dallas brand color palette.
}

