% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_color_callier}
\alias{scale_color_callier}
\title{Create a ggplot Color Scale Using Callier Center Colors}
\usage{
scale_color_callier(scheme, steps, ...)
}
\arguments{
\item{scheme}{A character vector that must be matched by
\code{'qual'} (qualitative palette), \code{'seq'} (sequential palette), or
\code{'div'} (diverging palette).}

\item{steps}{An integer, the number of colors in the scale.}

\item{...}{optional arguments passed to \code{CallierSequential()},
\code{CallierDiverging()}, or \code{scale_color_manual()}.}
}
\description{
\code{scale_color_callier()} creates a discrete color scale using Callier
Center colors, which can be added to a \code{ggplot} object.
}
\seealso{
\code{\link{CallierQualitative}}, \code{\link{CallierSequential}},
  and \code{\link{CallierDiverging}} for the color palette constructors;
  \code{\link{scale_color_manual}} and \code{\link{discrete_scale}} for the
  ggplot scale constructors.
}

