% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seedlingGreen}
\alias{#C9DD03}
\alias{#c9dd03}
\alias{seedlingGreen}
\title{Seedling Green}
\format{A character string, \code{'#C9DD03'}.}
\source{
Callier Center Graphic Standards Manual,
  UT Dallas Brand Standards (\url{https://www.utdallas.edu/brand/palette/})
}
\usage{
seedlingGreen
}
\description{
The hexadecimal code that specifies Pantone 381 ('seedling green'), which
corresponds to 201-221-3 in the RGB (red-green-blue) color model.
}
\details{
Seedling green is one of the three tertiary Callier Center colors
(eco green, sapling green, seedling green) that are preferred as accents to
complement the primary Callier Center colors in print or electronic formats.
The tertiary Callier Center colors are the three green-hued colors from the
UT Dallas brand color palette.
}
\keyword{datasets}

