% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{skyBlue}
\alias{#5BC6EB}
\alias{#5bc6eb}
\alias{skyBlue}
\title{Sky Blue}
\format{A character string, \code{'#5BC6EB'}.}
\source{
Callier Center Graphic Standards Manual
}
\usage{
skyBlue
}
\description{
The hexadecimal code that specifies Pantone 2985 blue ('sky blue'),
which corresponds to 91-198-232 in the RGB (red-green-blue) color model.
}
\details{
Sky blue is one of the three secondary Callier Center colors
(spark orange, stratos blue, sky blue) that are preferred as accents to
complement the primary Callier Center colors in print or electronic formats.
}
\keyword{datasets}

