% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standards-colors.R
\name{solarOrange}
\alias{#E98300}
\alias{#e98300}
\alias{solarOrange}
\title{Solar Orange}
\source{
Callier Center Graphic Standards Manual,
  UT Dallas Brand Standards (\url{https://www.utdallas.edu/brand/palette/})
}
\usage{
solarOrange()
}
\value{
A character string, \code{'#E98300'}.
}
\description{
The hexadecimal code that specifies Pantone 144 ('solar orange'),
which corresponds to 233-131-0 in the RGB (red-green-blue) color model.
}
\details{
Solar orange is one of the three primary Callier Center colors
(solar orange, space blue, callier gray), which should be used when depicting
Callier Center colors in print or electronic formats. Of the primary
Callier Center colors, solar orange has first precedence, space blue has
second, and callier gray has third. Hence, one-color images should use
solar orange; two-color images should use solar orange and space blue; and
three-color images should use solar orange, space blue, and callier gray.
}

