% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standards-colors.R
\name{spaceBlue}
\alias{#0039A6}
\alias{#0039a6}
\alias{spaceBlue}
\title{Space Blue}
\source{
Callier Center Graphic Standards Manual,
  UT Dallas Brand Standards (\url{https://www.utdallas.edu/brand/palette/})
}
\usage{
spaceBlue()
}
\value{
A character string, \code{'#0039A6'}.
}
\description{
The hexadecimal code that specifies Pantone 286 ('space blue'),
which corresponds to 0-57-166 in the RGB (red-green-blue) color model.
}
\details{
Space blue is one of the three primary Callier Center colors
(solar orange, space blue, callier gray), which should be used when depicting
Callier Center colors in print or electronic formats. Of the primary
Callier Center colors, solar orange has first precedence, space blue has
second, and callier gray has third. Hence, one-color images should use
solar orange; two-color images should use solar orange and space blue; and
three-color images should use solar orange, space blue, and callier gray.
}

