% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{CallierQualitative}
\alias{CallierQualitative}
\title{Create Qualitative Color Palettes with Callier Center Colors}
\usage{
CallierQualitative(steps = 3)
}
\arguments{
\item{steps}{An integer, the number of colors in the palette. Default is 3.}
}
\value{
A character vector of hexadecimal codes for the Callier Center colors.
  If \code{steps} is greater than the number of Callier Center colors (6),
  then \code{NULL} is returned; hence, this palette cannot be used for
  qualitative data with more than 6 levels.
}
\description{
\code{CallierQualitative()} creates a qualitative color palette based on
the primary and secondary Callier Center colors, in the following order of
precedence: solar orange, space blue, callier gray, spark orange,
stratos blue, sky blue.
}
\details{
A qualitative color palette does not imply magnitude differences between
levels in the data. As such, a qualitative palette is best suited to
representing nominal or categorical data.

A qualitative color palette is most effective when hues are used to signal
visual differences between levels in the data; hence, a palette that
uses only the three primary Callier Center colors is thus quite effective.
But the effectiveness of the Callier Center colors as a qualitative palette
diminishes when the secondary Callier Center colors are used: the secondary
colors all repeat hues (orange or blue) that are already represented in the
primary Callier Center colors.
}
\seealso{
\code{\link{CallierSequential}}, \code{\link{CallierDiverging}}
}

