% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-chart.R
\name{ColorChart}
\alias{ColorChart}
\title{Print a Color Chart for a Palette}
\usage{
ColorChart(palette = NULL, colorblind = TRUE, chipsize = 25)
}
\arguments{
\item{palette}{\code{NULL} (default) or a character vector, whose entries
are hexadecimal codes for values in the RGB (red-green-blue) color model.}

\item{colorblind}{Logical that determines whether simulations of
color blindness (deuteranopia, protanopia, tritanopia) are shown
(\code{TRUE}) or not (\code{FALSE)}) along with the trichromat color
palette. Default is \code{TRUE}.}

\item{chipsize}{Numeric that is passed to \code{geom_point(size = )}.
Determines the size of the color chips in the chart. Default is 25.}
}
\value{
A \code{ggplot} object with the following aesthetics:
  colors are ordered as they occur in \code{palette} and mapped to the
  x-dimension, with the x-axis labeled by hexadecimal codes;
  vision is mapped to the y-dimension, with the y-axis labeled by type of
  vision (trichromat, deuteranopia, protanopia, tritanopia).
}
\description{
\code{ColorChart()} creates a visual reference of a color palette as it
would appear to trichromat viewers and to viewers with three types of
color blindness:
deuteranopia (red-green color blindness; 1\% of males),
protanopia (red-green color blindness; 1\% of males), and
tritanopia (blue-yellow color blindness; < 1\% of males and females).
}
\details{
If \code{palette} is \code{NULL}, then a color chart for the six Callier
Center colors (solar orange, space blue, callier gray, spark orange,
stratos blue, sky blue) is created, and the x-axis is labeled with the
color names rather than their hexadecimal codes.

The \code{dichromat} package is used to simulate color blindness.
}

