% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standards-colors.R
\name{skyBlue}
\alias{#5BC6EB}
\alias{#5bc6eb}
\alias{skyBlue}
\title{Sky Blue}
\source{
Callier Center Graphic Standards Manual,
  UT Dallas Brand Standards (\url{https://www.utdallas.edu/brand/palette/})
}
\usage{
skyBlue()
}
\value{
A character string, \code{'#5BC6EB'}.
}
\description{
The hexadecimal code that specifies Pantone 2985 ('sky blue'),
which corresponds to 91-198-232 in the RGB (red-green-blue) color model.
}
\details{
Sky blue is one of the three secondary Callier Center colors
(spark orange, stratos blue, sky blue) that are preferred as accents to
complement the primary Callier Center colors in print or electronic formats.
}

