% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standards-colors.R
\name{sparkOrange}
\alias{#FFB612}
\alias{#ffb612}
\alias{sparkOrange}
\title{Spark Orange}
\source{
Callier Center Graphic Standards Manual,
  UT Dallas Brand Standards (\url{https://www.utdallas.edu/brand/palette/})
}
\usage{
sparkOrange()
}
\value{
A character string, \code{'#FFB612'}.
}
\description{
The hexadecimal code that specifies Pantone 1235 ('spark orange'),
which corresponds to 255-182-18 in the RGB (red-green-blue) color model.
}
\details{
Spark orange is one of the three secondary Callier Center colors
(spark orange, stratos blue, sky blue) that are preferred as accents to
complement the primary Callier Center colors in print or electronic formats.
}

