% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stratosBlue}
\alias{#00A1DE}
\alias{#00a1de}
\alias{stratosBlue}
\title{Stratos Blue}
\format{A character string, \code{'#00A1DE'}.}
\source{
Callier Center Graphic Standards Manual,
  UT Dallas Brand Standards (\url{https://www.utdallas.edu/brand/palette/})
}
\usage{
stratosBlue
}
\description{
The hexadecimal code that specifies Pantone 299 ('stratos blue'),
which corresponds to 0-161-222 in the RGB (red-green-blue) color model.
}
\details{
Stratos blue is one of the three secondary Callier Center colors
(spark orange, stratos blue, sky blue) that are preferred as accents to
complement the primary Callier Center colors in print or electronic formats.
}
\keyword{datasets}

